Se agrego la funcion de registrarse y logearse utilizando jwt y hashes para el password asi como la restriccion de los endpoints

+++ volcanes-api/Models/DTOs/VolcanActualizarDTO.cs
@@ -5,7 +5,7 @@ public class VolcanActualizarDTO
     public int Id { get; set; }
     public string Nombre { get; set; }
     public string Descripcion { get; set; }
-    public double Altura { get; set; }
+    public float Altura { get; set; }
     public string Ubicacion { get; set; }
     public string Ecosistema { get; set; }
 }
